+++ src/exception/io/io.ts
@@ -1,0 +1,8 @@
+import { Exception } from 'exception'
+
+/**
+ * Generic Input/Output exception
+ */
+export class IOException<T extends any[] = []> extends Exception<T> {
+
+}