+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### 0.15.0
 `+` New Async methods for `Signal`  
 `+` New `Exception` class  
+`*` Bunch of updates for the `I18nManager` and `LanguageDictionary`  
 `*` Refactor `CancellableSignal` to `PrefixedSignal`, make some behavioral changes to it  
 `*` Update CHANGELOG format